requirements.txt
Django>=5.0,<6.0
django-ninja>=0.20.0
